requirements.txt
altair==4.2.0
folium==0.13.0
geopandas
matplotlib==3.5.2
numpy==1.21.5
pandas
Pillow==9.4.0
plotly==5.9.0
pyproj
pyshp==2.3.1
python_dateutil==2.8.2
scipy
seaborn==0.11.2
streamlit==1.14.0
streamlit_metrics==0.1.0
streamlit_option_menu==0.3.2
openpyxl
